requirements.txt
Flask
google-generativeai
google-cloud-texttospeech
python-dotenv





